style: move pages and layout pages out of components folder
