+++ src/layouts/WorkspaceLayout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import WorkspaceBottomNav from './WorkspaceBottomNav.jsx';
+import WorkspaceBottomNav from '../components/WorkspaceBottomNav.jsx';
 
 const WorkspaceLayout = () => (
   <>
