+++ src/pages/Statistics.jsx
